Include test outputs in TestResult.
This makes it possible to retrieve them without doing I/O while emitting BEP events (*). I considered making the test outputs part of TestResultData (as suggested by the now deleted TODO in TestResult) but didn't like the resulting amount of proto conversion logic, and it's unclear that it fits with future plans.

Also rename FailedAttemptResult to ProcessedAttemptResult, as after this CL it sometimes represents the result of the final, and possibly succeeding, test attempt.

Progress on #17911.

(*) It doesn't currently save much, because we need I/O anyway to compute digests for the BEP. However, it's incremental progress towards a better architecture (i.e. in the future we could pass artifacts and their metadata around, instead of just filenames).

PiperOrigin-RevId: 673364262
Change-Id: I9266b7708d90ac252336c813a24c166712a699c7